Question shuffling
Code created to shuffle questions and answers in a multiple choice exam written in LaTeX for a class I was teacher assistant. Format I used for the exam is:
\item Question\\
\vspace{-6mm}
\begin{enumerate}
\itemsep-4mm
\item Choice 1 \\
\item Choice 2 \\
\item Choice 3 \\
\item Choice 4 \\
\item Choice 5 \\
\end{enumerate}

so the code can be changed if you are using different formatting. It was required by the teachers that the last question wasn't shuffled with the others (but the choices must be), so you can comment also that part of the code